Split aggregators out into separate files.
