+++ src/ad.js
@@ -5,4 +5,14 @@ ad.isTape = function(obj) {
   return _.has(obj, 'primal');
 };
 
+// Recursively untapify objects. ad.js already does this for arrays,
+// here we extend that to other objects.
+ad.deepUntapify = function(x) {
+  if (_.isObject(x) && !_.isArray(x) && !ad.isTape(x)) {
+    return _.mapObject(x, ad.deepUntapify);
+  } else {
+    return ad.untapify(x);
+  }
+};
+
 module.exports = ad;
