+++ src/inference/asyncpf.js
@@ -8,7 +8,7 @@
 
 var _ = require('underscore');
 var util = require('../util');
-var Histogram = require('../aggregation').Histogram;
+var Histogram = require('../aggregation/histogram');
 
 module.exports = function(env) {
